[FEAT] SSE 알림 기능 구현, notification 구현

+++ src/main/java/com/crewing/member/dto/MemberListResponse.java
@@ -10,6 +10,6 @@ import java.util.List;
 public class MemberListResponse {
     private int pageNum;
     private int pageSize;
-    private int totalCnt;
+    private Long totalCnt;
     private List<MemberInfoResponse> members;
 }
